Separated actions and reducers into separate folders
