+++ autoscheduler/frontend/src/redux/reducers/availabilityMode.ts
@@ -2,7 +2,7 @@
  * Sets whether the user is selecting busy time or not.
  * In the future, this could also support preferred times
  */
-import { AvailabilityType } from '../types/Availability';
+import { AvailabilityType } from '../../types/Availability';
 
 // action type string
 export const SET_AVAILABILITY_MODE = 'SET_AVAILABILITY_MODE';
