+++ autoscheduler/frontend/src/redux/reducers/courseCards.ts
@@ -2,7 +2,7 @@
  * Stores all course cards that the user uses to select courses to include
  * in generated schedules
  */
-import { CourseCardOptions, CourseCardArray, CustomizationLevel } from '../types/CourseCardOptions';
+import { CourseCardOptions, CourseCardArray, CustomizationLevel } from '../../types/CourseCardOptions';
 
 // action type strings
 export const ADD_COURSE_CARD = 'ADD_COURSE_CARD';
