+++ autoscheduler/frontend/src/tests/types/Availability.test.ts
@@ -1,7 +1,7 @@
 import { createStore } from 'redux';
 import Availability, { roundUpAvailability, AvailabilityArgs, AvailabilityType } from '../../types/Availability';
 import autoSchedulerReducer from '../../redux/reducer';
-import { addAvailability } from '../../redux/actions';
+import { addAvailability } from '../../redux/actions/availability';
 import DayOfWeek from '../../types/DayOfWeek';
 
 /**
